Dynamic server branding support
The Online server product name is now
dynamically pulled from the server, once a
valid URL is provided. To avoids using
a hard-coded name. After the initial
discovery of the server capabilities, the
product-name is cached.

However we allow for changing the server
(or, upon updating the server the product
name may have been changed) by clearing
the cache from the plugin settings page.
Simply loading the settings page will
invalidate the cache, if a URL is already
set and is valid, and will pull the new
name.

The relevant strings displayed in the UI,
including errors, have been updated to use
the product name provided by the server
with proper fallback.