Support per-document options
This adds support for per-document
options, such as enabling/disabling
printing, copying, exporting, etc.
and adds the ability to set a watermark
and other security settings.

These option/settings can be modified
by the document owner. The document locking
logic is also included as a form option to
make it homogeneous with the others.

The options are hidden by default, in a collapsed
section. They are never shown to user without
the ability to modify them.